fix: change all avg fields from int to float Closes #6

+++ src/api/schemas/review-store.admin.schema.ts
@@ -19,7 +19,7 @@ export const reviewStoreAdminApiExtension = gql`
   }
   extend type Query {
     "Get the average of review store"
-    avgReviewStore: Int
+    avgReviewStore: Float
 
     "Get the review store"
     reviewStore(id: ID!): ReviewStore
